+++ cmd/node/main.go
@@ -4,6 +4,8 @@
 package main
 
 import (
+	"context"
+	"errors"
 	"fmt"
 	"os"
 	"os/signal"
@@ -15,22 +17,23 @@ import (
 // main executes the antler-node command.
 func main() {
 	if len(os.Args) != 2 {
-		fmt.Fprintf(os.Stderr, "%s: exactly one argument required (node ID)\n")
+		fmt.Fprintf(os.Stderr, "%s: exactly one argument required (node ID)\n",
+			os.Args[0])
 		fmt.Fprintf(os.Stderr, "usage: %s <node ID>\n", os.Args[0])
 		os.Exit(1)
 	}
 	n := os.Args[1]
-	ct := node.NewControl()
-	defer ct.Stop()
+	c, x := context.WithCancelCause(context.Background())
+	defer x(nil)
 	i := make(chan os.Signal, 1)
 	signal.Notify(i, os.Interrupt, syscall.SIGTERM)
 	go func() {
 		s := <-i
 		fmt.Fprintf(os.Stderr, "%s, canceling\n", s)
-		ct.Cancel(s.String())
+		x(errors.New(s.String()))
 	}()
-	c := node.StdioConn()
-	if err := node.Serve(n, ct, c); err != nil {
+	o := node.StdioConn()
+	if err := node.Serve(c, n, o); err != nil {
 		fmt.Fprintf(os.Stderr, "node exiting with status 1: %s\n", err)
 		os.Exit(1)
 	}